refactor: _components 폴더 추가 및 정리
