+++ src/app/home/_components/banner-options.tsx
