+++ src/app/home/_components/rectangle-poster.tsx
