+++ src/app/home/_components/rectangle-viewer.tsx
